+++ Src/Library/Binder/BinderExtensions.cs
@@ -10,6 +10,12 @@ namespace FastEndpoints;
 
 internal static class BinderExtensions
 {
+    internal static IEnumerable<PropertyInfo> BindableProps(this Type t)
+    {
+        return t.GetProperties(BindingFlags.Public | BindingFlags.Instance | BindingFlags.FlattenHierarchy)
+                .Where(p => p.GetSetMethod()?.IsPublic is true && p.GetGetMethod()?.IsPublic is true);
+    }
+
     internal static Func<object, object> GetterForProp(this Type source, string propertyName)
     {
         //(object parent, object returnVal) => ((object)((TParent)parent).property);