+++ Src/Library/changelog.md
@@ -3,5 +3,6 @@
 ### IMPROVEMENTS
 - remove readonly request dto properties from swagger doc #306
 - capitalize first letter of swagger auto tags based on path segment
+- default request binder will ignore dto props without public getter/setter
 
 ### FIXES